Fix sync bug
We were previously dispatching actions to each slice in response to a
sync dispatch. The problem here is that sync runs on every change to the
state, so it was running on every dispatch.

If the first dispatch didn't bring the states into alignment, syncyjson
would dispatch it again in response, causing an infinite loop.